The toClockwise() code that was added to the XMLControl classes was incorrect - this was only valid for north-up projections.  Removed this entirely - it is up to the caller to order the corners properly.
Made the corners more explicit by creating a LatLonCorners struct which spells out the four corners.  This class also has a getCorner() method which can be used to cycle through the four corners where this is useful (could use operator[] instead / in addition to this but was trying to examine all the code where the vector had been used to make sure the caller was always ordering the corners properly).  Replaced the CornerIndex enums with some static constants in the LatLonCorners class (this appears to have only been used in the six-samples anyhow).

Fixed a bug in NITFReadControl where the corners that went into the NITFSegmentInfo were assigned incorrectly (looked like a cut and paste error).  Also, in parseFootprint(), the index was parsed but not actually used when inserting into the corners, so updated the code to actually use the index (in reality I think the corners are always specified in the XML with indices 1-4 in order so this didn't matter).

test_round_trip was deleting the container when it shouldn't have been and had a memory leak in that it wasn't deleting the command-line options, so wrapped this in an auto_ptr.

git-svn-id: https://wush.net/svn/mdaus-oss/trunk@291 9d772353-a6c9-4ee9-b2cc-df47a7649cd5