Squashed 'externals/coda-oss/' changes from a12bbff..00439a0
00439a0 Merge pull request #44 from mdaus/vs2015_fix
eae4787 Removing logic that overrode cmd_and_log() as this was causing problems with VS 2015 Express (it does not have all valid paths set in LIB, and that's ok, but the old check was trying to find bad configurations and avoid them).  See comments in build.py for more details.  I have a theory that this was what was also causing waf to not like to use the x64 target with some of our VS Pro installations, so I'm also reverting the temporary hack where we'd preferred x86_amd64 over it.  If versions of VS Express / Windows SDK prior to 2012 aren't working, this is a commit to look at.

git-subtree-dir: externals/coda-oss
git-subtree-split: 00439a00af8964b68b6816350a05957232217a5e