Squashed 'externals/coda-oss/' changes from 740accd..1335087
1335087 Merge pull request #46 from mdaus/TypesRowColSqrtFixForWindows
56f624a Clarifying comment and simplifying by only doing a single cast
939030d formatting
f829ccc Visual Studio 2010 can't resolve the sqrt for size_t or sys::SSize_T for the python bindings. This is not a heavily used method, and because types does not depend on sys for Conf.h, we do an explicit cast to double.

git-subtree-dir: externals/coda-oss
git-subtree-split: 1335087e6a85e3c753584446ab8efaf6355e1d87