Squashed 'externals/coda-oss/' changes from 2fc9ea7..37526a4
37526a4 Merge pull request #61 from mdaus/python_ld_update
5772485 Fixed OSX -bundle removal code
39820a3 Removed conflicting -bundle from osx linkflags
05df427 More changes for compiling on OSX
be7d392 Modified to hopefully support OSX's ld
eee65f8 Hopefully adds support for ld >2.17
b92bf61 Merge pull request #57 from mdaus/SysExecTerminate
4cd1cb4 adding static keyword
1d1d748 updating constants capitalization
c105967 removing the brackets and unnecessary break
9b60823 Updates to address all review updates except the curly braces in the case statement. Waiting for further clarification there
4a61eef updating the pipe naming
6d79b5e made the executive decision std::cin will be passed via the parent's std::cin. This is a simplification that should be useful in this context. This simplifies the class a bit, and provides all the functionality required on known applications.
00bb171 we can't close the pipes sent into the subproces or it will close early
f41b380 adding the ability to add arguments to the processing
a403721 updated the naming based on a #define from UNICODE
4786fde additional cleanup and closure of pipes
05e039c removing this debugging job
a4cd2da Updated to use the non-deprectated calls, and long buffering was the problem before. Setting the buffer size to 1 solves this issue.
a02211f preliminary check-in. This is working when I right to the parent's stdout pipe, but not in the FILE* context.
4332dc2 Updates to get the Linux side matching what I did for FSG. One character makes all the difference
16ce33d updating to work on Linux
a9da04a This is a branch for terminating the child process swiftly without discretion. For this we must create our own popen command, which does not internally perform a wait() command. This will allow the subprocess to schedule, but not require the completion of a job in order to terminate. This helps in the case when the parent process is terminated, and subprocess continues on in a zombie state until its underlying process legitimately completes. Now when the parent dies, so does the child.

git-subtree-dir: externals/coda-oss
git-subtree-split: 37526a45be6e648ad912f081e40acff6a460fb2b